requirements.txt
annotated-types==0.7.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
Flask==3.1.2
garminconnect==0.2.36
garth==0.5.19
google-cloud-storage==2.18.2
google-generativeai==0.8.3
gunicorn==23.0.0
python-dotenv==1.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
oauthlib==3.3.1
pydantic==2.12.5
pydantic_core==2.41.5
requests==2.32.5
requests-oauthlib==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.1
Werkzeug==3.1.4
